Race condition for detecting the real command name of the spawned ffmpeg process.
We therefore hardcode the command name for logging purposes.
It is a bit of a hack, but we already know what it should be and the
process spawned for running ffmpeg actually changes command name twice:
basename $0 (due to coproc), nohup and finally ffmpeg
To ilustrate this, add the following code at line 385, log stderr of
this script and run it under load or when the system is booting.

deadline=$((SECONDS+2))
while [ "$(ps -p ${FFPID} -o comm=)" != ffmpeg ]; do
    echo "$(ps -p ${FFPID} -o comm=)" >&2
    [ $SECONDS -eq $deadline ] && break
done